html/semantics/embedded-content-0/media-elements/canPlayType_application_octet_stream_with_codecs_2.html

<!doctype html>
<html>
 <head>
  <title>{audio,video}.canPlayType() - can never play application/octet-stream with a codecs parameter</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
  <p><a href="http://dev.w3.org/html5/spec/Overview.html#a-type-that-the-user-agent-knows-it-cannot-render">spec reference</a></p>
  <audio id="a">
  </audio>
  <video id="v">
  </video>
  <div id="log"></div>
  <script>
test(function() {
 assert_equals(
  document.getElementById("a").canPlayType('application/octet-stream; codecs="mp4a.40.2"'),
  "",
  "audioElement.canPlayType('application/octet-stream; codecs=...') should always return the empty string");
}, "audio.canPlayType() - can never play application/octet-stream with a codecs parameter");
  
test(function() {
 assert_equals(
  document.getElementById("v").canPlayType('application/octet-stream; codecs="theora, vorbis"'),
  "",
  "videoElement.canPlayType('application/octet-stream; codecs=...') should always return the empty string");
}, "video.canPlayType() - can never play application/octet-stream with a codecs parameter");
  </script>
 </body>
</html>
